ReportGenerationFunction/CloudFunction/userlogoutactivity/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-firestore==2.0.2
datetime